First working version of an unbundled openjdk 8 cartridge.
It sets an environment variable called OPENSHIFT_OPENJDK8_JAVA_HOME
which can be consumed by Java web_framework cartridges.

The cartridge is called "java-openjdk-1.8.0" and may be added as an
add-on to e.g. jbossews-2.0. Example app which uses a modified
jbossews-2.0 cartridge in order to enable Java 8 support:

$ rhc app create testapp jbossews-2.0 openjdk8

Then in order to switch to Java 8 use the following:

$ cd testapp
$ git mv .openshift/markers/java{7,8}
$ git commit -m "Using Java 8"
$ git push origin master